requirements.txt
akshare>=1.12.0
requests>=2.31.0
schedule>=1.2.0
python-dotenv>=1.0.0
openai>=1.0.0
pandas>=2.0.0
flask>=2.3.0
dingtalk-stream>=0.9.0
tushare>=1.2.0
yfinance>=0.2.0
baostock>=0.8.0
